docs: `useUiNotification` composable (M2-397) (#904)
